requirements.txt
gradio==4.44.0
ollama==0.3.3
python-dotenv==1.0.0
mcp==1.10.1 